Adjusted image heights
Setting both width and height to the same value results in odd-looking
images with incorrect proportions.

When no height or with is given, the browser renders the image in its
original size.

With just one of the two (width or heigh) specified, the browser
determines the other from the images original proportion.

Setting a width and height greater than the images original size makes
the image appear grainy. This is because the image is defined by its
individual pixels, which are finite.